reformmat the codes for utility computation
