+++ PETWorks/ambiguity.py
@@ -1,5 +1,3 @@
-from typing import Dict, List
-
 from PETWorks.arx import (
     Data,
     loadDataFromCsv,
